+++ lib/linkedin_login.dart
@@ -4,7 +4,6 @@ export 'package:linkedin_login/src/model/linked_in_user_model.dart';
 
 export 'package:linkedin_login/src/utils/widgets/linked_in_buttons.dart';
 export 'package:linkedin_login/src/wrappers/authorization_code_response.dart';
-export 'package:linkedin_login/src/wrappers/linked_in_error_object.dart';
 export 'package:linkedin_login/src/wrappers/linked_in_token_object.dart';
 export 'package:linkedin_login/src/client/linked_in_user_widget.dart';
 export 'package:linkedin_login/src/server/linked_in_auth_code_widget.dart';